Moved Trip components to trips folder
