+++ TripBorderFrontend/src/components/trips/ClickMarker.jsx
