+++ TripBorderFrontend/src/components/trips/DirectionLayer.jsx
