+++ TripBorderFrontend/src/components/trips/ProximityMarkers.jsx
@@ -7,9 +7,9 @@ import {
   setIsShowingAddtionalPopUp,
   setIsShowingOnlySelectedPOI,
   setSelectedPOILonLat
-} from '../redux/reducers/mapReducer';
-import { FourSquareResponsePropTypes } from '../constants/fourSquarePropTypes';
-import CustomButton from './CustomButton';
+} from '../../redux/reducers/mapReducer';
+import { FourSquareResponsePropTypes } from '../../constants/fourSquarePropTypes';
+import CustomButton from '../CustomButton';
 
 // eslint-disable-next-line react/prop-types
 export default function ProximityMarkers({ data, getPOIPhotosQueryTrigger }) {
